# backend/docs/swagger.yaml
basePath: /
definitions:
  handler.BetListResponse:
    properties:
      bets:
        items: {}
        type: array
    type: object
  handler.ChangeDeciderRequest:
    properties:
      decider_id:
        example: 665b1a2b3c4d5e6f7a8b9c0d
        type: string
    type: object
  handler.CreateBetRequest:
    properties:
      decider_id:
        example: 665b1a2b3c4d5e6f7a8b9c0d
        type: string
      description:
        example: Sam has lived in Provo for 2 years...
        type: string
      end_date:
        example: "2026-04-01T23:59:59Z"
        type: string
      opening_wager:
        $ref: '#/definitions/handler.OpeningWagerRequest'
      title:
        example: Sam won't hike the Y before spring
        type: string
    type: object
  handler.CreateGroupRequest:
    properties:
      description:
        example: Our main friend group.
        type: string
      name:
        example: The Squad
        type: string
    type: object
...
  handler.ErrorDetail:
    properties:
      code:
        example: NOT_FOUND
        type: string
      message:
        example: not found
        type: string
    type: object
  handler.ErrorResponse:
    properties:
      error:
        $ref: '#/definitions/handler.ErrorDetail'
    type: object
  handler.GroupListResponse:
    properties:
      groups:
        items: {}
        type: array
    type: object
...
  handler.JoinByCodeRequest:
    properties:
      invite_code:
        example: XK9F2M
        type: string
    type: object
...
  handler.OpeningWagerRequest:
    properties:
      amount:
        example: 100
        type: integer
      side:
        enum:
        - FOR
        - AGAINST
        example: FOR
        type: string
    type: object
  handler.PayoutEntry:
    properties:
      amount:
        example: 175
        type: integer
      name:
        example: Sam
        type: string
      net:
        example: 75
        type: integer
      user_id:
        example: 665a1a2b3c4d5e6f7a8b9c0d
        type: string
    type: object
  handler.PlaceWagerRequest:
    properties:
      amount:
        example: 200
        type: integer
      side:
        enum:
        - FOR
        - AGAINST
        example: AGAINST
        type: string
    type: object
  handler.ResolveRequest:
    properties:
      winning_side:
        enum:
        - FOR
        - AGAINST
        example: FOR
        type: string
    type: object
  handler.ResolveResponse:
    properties:
      id:
        example: 665f1a2b3c4d5e6f7a8b9c0d
        type: string
      payouts:
        items:
          $ref: '#/definitions/handler.PayoutEntry'
        type: array
      resolved_at:
        example: "2026-04-02T15:30:00Z"
        type: string
      status:
        example: RESOLVED
        type: string
      winning_side:
        example: FOR
        type: string
    type: object
...
  model.Bet:
    properties:
      created_at:
        type: string
      creator_id:
        type: string
      decider_id:
        type: string
      description:
        type: string
      end_date:
        type: string
      group_id:
        type: string
      id:
        type: string
      resolved_at:
        type: string
      status:
        type: string
      title:
        type: string
      wagers:
        items:
          $ref: '#/definitions/model.Wager'
        type: array
      winning_side:
        type: string
    type: object
  model.Group:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      description:
        type: string
      id:
        type: string
      invite_code:
        type: string
      members:
        items:
          $ref: '#/definitions/model.Member'
        type: array
      name:
        type: string
    type: object
...
  model.Member:
    properties:
      joined_at:
        type: string
      role:
        type: string
      user_id:
        type: string
    type: object
...
  model.Wager:
    properties:
      amount:
        type: integer
      id:
        type: string
      placed_at:
        type: string
      side:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Social prediction betting API — create groups, place bets, resolve
    outcomes.
  title: youwont API
  version: "1.0"
paths:
  /bets/{id}:
    get:
      description: Returns full bet detail including all wagers. Requires membership
        in the bet's group.
      parameters:
      - description: Bet ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Bet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get bet detail
      tags:
      - bets
  /bets/{id}/decider:
    put:
      consumes:
      - application/json
      description: Change the decider on an open bet. Only the bet creator can do
        this. New decider must not have wagered.
      parameters:
      - description: Bet ID
        in: path
        name: id
        required: true
        type: string
      - description: New decider
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.ChangeDeciderRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Bet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Change bet decider
      tags:
      - bets
  /bets/{id}/resolve:
    post:
      consumes:
      - application/json
      description: Resolve an open bet by picking the winning side. Only the decider
        can resolve. Both sides must have wagers. Points are distributed proportionally
        and notifications are sent to all participants.
      parameters:
      - description: Bet ID
        in: path
        name: id
        required: true
        type: string
      - description: Winning side
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.ResolveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ResolveResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Resolve a bet
      tags:
      - bets
  /bets/{id}/wagers:
    post:
      consumes:
      - application/json
      description: Place a wager on an open bet. One wager per user per bet. Decider
        cannot wager. Min 10 pts. Points deducted immediately.
      parameters:
      - description: Bet ID
        in: path
        name: id
        required: true
        type: string
      - description: Wager details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.PlaceWagerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Bet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Place a wager
      tags:
      - bets
  /groups:
    get:
      description: Returns all groups the authenticated user is a member of.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.GroupListResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List my groups
      tags:
      - groups
    post:
      consumes:
      - application/json
      description: Creates a new group with the authenticated user as ADMIN. Auto-generates
        an invite code.
      parameters:
      - description: Group details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateGroupRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Group'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a group
      tags:
      - groups
  /groups/{id}:
    get:
      description: Returns group info including members. Requires group membership.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Group'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get group detail
      tags:
      - groups
  /groups/{id}/bets:
    get:
      description: Returns all bets in a group, optionally filtered by status. Requires
        group membership.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Filter by status
        enum:
        - OPEN
        - RESOLVED
        - CANCELED
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.BetListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List bets in a group
      tags:
      - bets
    post:
      consumes:
      - application/json
      description: Create a new bet with an opening wager. Creator cannot be the decider.
        Points are deducted immediately. Notifies group members.
      parameters:
      - description: Group ID
        in: path
        name: id
        required: true
        type: string
      - description: Bet details with opening wager
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.CreateBetRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Bet'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a bet
      tags:
      - bets
...
  /groups/join:
    post:
      consumes:
      - application/json
      description: Join a group using its invite code. Fallback for when in-app invites
        aren't used.
      parameters:
      - description: Invite code
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.JoinByCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Group'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/handler.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Join group via invite code
      tags:
      - groups
...
securityDefinitions:
  BearerAuth:
    description: Enter your Supabase JWT token with the "Bearer " prefix.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"